Convert PG/AMQP funcs into structs
