+++ event-store/src/adapters/emitter/mod.rs
@@ -1,0 +1,3 @@
+mod amqp;
+
+pub use self::amqp::AmqpEmitterAdapter;
