+++ event-store/src/adapters/store/mod.rs
@@ -1,0 +1,3 @@
+mod pg;
+
+pub use self::pg::{PgQuery, PgStoreAdapter};
