+++ event-store/src/lib.rs
@@ -6,17 +6,12 @@
 // only needed to manually implement a std future:
 #![feature(arbitrary_self_types)]
 
-// Bring tokio's shimmed await!() into scope
-// #[macro_use]
-// extern crate tokio;
-
+pub mod adapters;
 pub mod aggregator;
-pub mod amqp;
 pub mod event;
 pub mod event_context;
 pub mod event_handler;
 pub mod event_replay;
-pub mod pg;
 pub mod store;
 pub mod store_query;
 pub mod subscribable_store;
@@ -46,11 +41,9 @@ pub fn forward<I, E>(
 }
 
 pub use crate::aggregator::*;
-pub use crate::amqp::*;
 pub use crate::event::Event;
 pub use crate::event_handler::*;
 pub use crate::event_replay::*;
-pub use crate::pg::*;
 pub use crate::store::*;
 pub use crate::store_query::*;
 pub use crate::subscribable_store::*;
